refactor(streams): rename compat.py to binaryio.py and clean its docstrings
compat.py read as a grab-bag with provenance narration ("ported from DEV's
scattered helpers ..."). It is actually one cohesive concern — classify and
coerce arbitrary caller objects to the uniform BinaryIO the library relies on —
so give it an intent-revealing name and describe purpose, not origin.

- rename internal/streams/compat.py -> binaryio.py
- move fix_stream_start_position into slice.py (it builds a SlicingStream and
  needed a lazy import to dodge a cycle; co-locating removes the cycle)
- rewrite docstrings to explain what/why, dropping DEV references (provenance
  lives in PLAN.md / openspec / git history, not shipped code); keep the
  BinaryIOWrapper plain-delegation rationale, reworded as a decision
- update importers + ARCHITECTURE.md; rename test_streams_compat.py ->
  test_stream_primitives.py

No behavior change. pyrefly + ty + ruff clean; 136 passed, 4 skipped.
